+++ index.html
@@ -26,7 +26,7 @@
           <!--  box1 -->
           <div class="box1">
             <!-- form -->
-            <form id="form" method="post">
+            <form id="form">
               <!-- form title -->
               <input
                 type="text"
@@ -46,7 +46,7 @@
               <button
                 aria-label="Form save"
                 id="btnSave"
-                type="submit"
+                type="button"
                 class="btnSave"
               >
                 <!-- form save icon -->
@@ -69,7 +69,6 @@
       </div>
     </main>
     <!-- Main -->
-
     <!-- footer -->
     <footer id="bottomFooter">
       Copyright © 2024 | Web Developer: Juan David Carrillo
