+++ package.json
@@ -31,6 +31,7 @@
   "homepage": "https://github.com/juanprogram1/FolderTypeScript#readme",
   "devDependencies": {
     "@types/node": "20.13.0",
+    "@types/uuid": "^9.0.8",
     "@typescript-eslint/eslint-plugin": "7.11.0",
     "@typescript-eslint/parser": "7.11.0",
     "eslint": "8.57.0",
