--- src/loadView.ts
@@ -1,20 +1,0 @@
-import { handleForm as handleForm } from "./btnSave";
-import { showBox2form } from "./btnSave";
-
-// viewForm function to show the content of the form in the text-content-box
-/* function viewForm() {
-
-    const contexform = `
-      <div class="text-content-box">
-        <h3 class="titleh3">${titleData}</h3>
-        <p class="date-text">${dateData}</p>
-        <p class="text-paragraph">${textareaData}</p>
-      </div>
-    `;
-
-    // show the content of the form in the text-content-box
-    showBox2form(contexform);
-  }
-}
-
-document.addEventListener("DOMContentLoaded", viewForm); // viewForm function to show the content of the form in the text-content-box */
